+++ app/Console/Kernel.php
@@ -14,6 +14,7 @@ class Kernel extends ConsoleKernel
      */
     protected $commands = [
         Commands\Serve::class,
+        Commands\GenerateSitemap::class,
     ];
 
     /**
@@ -35,6 +36,21 @@ class Kernel extends ConsoleKernel
      */
     protected function commands()
     {
-        require base_path('routes/console.php');
+        require $this->app->basePath().'/routes/console.php';
+    }
+
+    /**
+     * Bootstrap the application for artisan commands.
+     *
+     * @return void
+     */
+    public function bootstrap()
+    {
+        // Bind the public path to public_html instead of public
+        $this->app->bind('path.public', function () {
+            return $this->app->basePath().'/public_html';
+        });
+
+        parent::bootstrap();
     }
 }