+++ app/Http/Requests/SendContactEmail.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class SendContactEmail extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'g-recaptcha-response' => 'bail|required|captcha',
+            'name'                 => 'bail|required|string',
+            'email'                => 'bail|required|email',
+            'message'              => 'required|string',
+        ];
+    }
+}